Move welcome fragment to fragment package
